Move anagram set directory to data directory
